<p style="text-align:center">
This app uses the NeuroSky MindWave (EEG sensor) with an Ardunio connected to a pulse sensor along with the Processing IDE because of its compatibility with the libraries. The MindWave was taken apart and reassembled with a durable sleeping mask for comfort. Its main function is to induce lucid dreams through playing musical triggers and binaural beats when the program detects the user to be in REM sleep, but it also functions excellently as a real-time EEG / biometric monitor, as the program records data and organizes it into spreadsheets and images for researching purposes.
<br><br>
Users will eventually be able to send data into the machine during their dreams by controlling eye blinking, and the software will be expanded with binaural beats, dynamic soundscapes, and visual cues (which would affect the environments of the dreams), and a refined, 3D printed mask will be created to house the components. Virtual reality environments (Through using Google Cardboard lenses and a smartphone) will immerse the user in order to stimulate a specific dream. I plan on creating an app for the device for Android / iOS, using Bluetooth to communicate between the two, and putting the final, refined product up on KickStarter.
<br><br> See the DreamGazer website where you can see all its details and features <a target="_blank"href="http://dreamgazer.me">here</a>.<br>